% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgread.R
\docType{package}
\name{dgread-package}
\alias{dgread}
\alias{dgread-package}
\title{dgread: Read Dynamic Group (dg/dgz) Data Files}
\description{
Fast reader for dg (dynamic group) and dgz (gzip-compressed) data files.
These files store collections of named arrays with support for nested and
ragged (variable-length) data, decompressed in memory (no temp file).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SheinbergLab/dgread}
  \item Report bugs at \url{https://github.com/SheinbergLab/dgread/issues}
}

}
\author{
\strong{Maintainer}: David Sheinberg \email{david_sheinberg@brown.edu}

Authors:
\itemize{
  \item David Sheinberg \email{david_sheinberg@brown.edu}
}

}
\keyword{internal}
